# backend/openapi.json
components:
  responses:
    DEFAULT_ERROR:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: Default error response
...
  schemas:
...
    CommonError:
      properties:
        errorCode:
          type: string
        message:
          type: string
      required:
      - errorCode
      - message
      type: object
...
    DocumentList:
      properties:
        documents:
          items:
            additionalProperties: {}
            type: object
          type: array
      type: object
    Error:
      properties:
        code:
          description: Error code
          type: integer
        errors:
          additionalProperties: {}
          description: Errors
          type: object
        message:
          description: Error message
          type: string
        status:
          description: Error name
          type: string
      type: object
...
    FileDownload:
      properties:
        file:
          description: The downloaded file
      type: object
...
info:
  title: My API
  version: v1
openapi: 3.0.2
paths:
...
  /api/download/{file_id}:
    get:
      description: "Args:\n    file_id (int): The ID of the file to download.\n\n\
        Returns:\n    The requested file as an attachment.\n\nRaises:\n    - 404:\
        \ If file not found in database or on server.\n    - 500: If there is a server\
        \ error."
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDownload'
          description: OK
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonError'
          description: Not Found
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonError'
          description: Internal Server Error
        default:
          $ref: '#/components/responses/DEFAULT_ERROR'
      summary: Download a submitted file by its ID.
      tags:
      - Submissions
    parameters:
    - in: path
      name: file_id
      required: true
      schema:
        minimum: 0
        type: integer
  /api/files/{team_id}:
    get:
      description: "Query Parameters:\n    - team_id (int): The ID of the team whose\
        \ submissions are being queried.\n\nReturns:\n    List of documents with their\
        \ metadata."
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentList'
          description: OK
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonError'
          description: Not Found
        default:
          $ref: '#/components/responses/DEFAULT_ERROR'
      summary: Get all files submitted by a specific team.
      tags:
      - Submissions
    parameters:
    - in: path
      name: team_id
      required: true
      schema:
        minimum: 0
        type: integer
...